Open Beta! Maybe.
As discussed, I cleaned up the focus tree for the military.  I also used two custom icons for the Nevada Path, one for the Southwest Commonwealth, and one for Seattle.   I cleaned up a few nits in some of the code as well.  The buildings will have to be checked to make sure that works.  But I think this is good enough for the open beta. (Let's charge bottlecaps).