Fixes/Changes
Move Nuke event to overwrite it so there isn't duplicates.
Change all lowercase not's, or's, and and's to upper case as required by macs for this game.
Fixed ROOT in the enc button files being being setup right.
Moved the units_deployed_to_overlord for the Lawless puppet to the autonomy file as required.